+++ src/main/resources/plugin.yml
@@ -1,7 +1,7 @@
 name: MapRights
 version: '${version}'
 main: com.archerwn.mapRights.MapRights
-api-version: '1.21'
+api-version: '1.20'
 author: ArcherWn
 description: A Minecraft plugin that allows players to sign maps, protecting their creations from unauthorized copying.
